backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sqlalchemy==2.0.15
mysql-connector-python==8.0.33
python-dotenv==1.0.0
